requirements.txt
backports.ssl-match-hostname==3.4.0.2
certifi==14.5.14
click==3.3
gunicorn==19.1.1
redis==2.10.3
rq==0.5.0
sockjs-tornado==1.0.1
tornado==4.0.2
